Fix: Commit transaction before destroying members
Previously, a Context object was first deleting all its members before
committing its transaction, which was sometimes leading to an invalid
state if a context member was bound to the transaction.

We noticed this error in score.sa.db, where the transaction failed due
to the database connection being closed (as the ctx member was already
destroyed).