Change: Member caching is now mandatory
We noticed that dynamic values are never actually used, probably because
they are just confusing from the point of view of the user of Context
objects. We found that another approach is used, instead, in these
cases: providing a function as context member.

>>> def value_generator():
...     return random.randint(1, 10)
...
>>> ctx_conf.register('dynamic_value', lambda ctx: value_generator)
>>> ctx_conf._finalize()
>>> ctx = ctx_conf.Context()
>>> ctx.dynamic_value()
4

This is now the only way to provide dynamic values as a context member.