Change: Context members can no longer be deleted
The semantics of deleting a member have always been a bit inconsistent:
Can a deleted member be accessed at a later point? Should the
constructor be invoked again? Shouldn't the destructor behave
differently when the member is deleted as when the context is being
cleaned up?

Context member's can no longer be deleted at all to avoid all this
ambiguity.